feat(smartwatch-app): Updated base package for smartwatch app
