+++ apps/smartwatch-app/.gitignore
