+++ apps/smartwatch-app/build.gradle.kts
@@ -1,0 +1,98 @@
+/*
+ * Copyright 2024 Marlonlom
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+@file:Suppress("DSL_SCOPE_VIOLATION")
+
+plugins {
+  alias(libs.plugins.android.application)
+  alias(libs.plugins.kotlin.android)
+}
+
+android {
+  namespace = "dev.marlonlom.apps.mocca.smartwatch"
+  compileSdk = 34
+
+  defaultConfig {
+    applicationId = "dev.marlonlom.apps.mocca.smartwatch"
+    minSdk = 30
+    //noinspection EditedTargetSdkVersion
+    targetSdk = 34
+    versionCode = 1
+    versionName = "1.0.0"
+
+    testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+
+    vectorDrawables {
+      useSupportLibrary = true
+    }
+
+  }
+
+  buildTypes {
+    release {
+      isMinifyEnabled = true
+      isShrinkResources = true
+      proguardFiles(
+        getDefaultProguardFile("proguard-android-optimize.txt"),
+        "proguard-rules.pro"
+      )
+    }
+  }
+  compileOptions {
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+  }
+  kotlinOptions {
+    jvmTarget = JavaVersion.VERSION_17.majorVersion
+  }
+  buildFeatures {
+    compose = true
+    buildConfig = true
+  }
+  composeOptions {
+    kotlinCompilerExtensionVersion = "1.5.8"
+  }
+  packaging {
+    resources {
+      excludes += "/META-INF/{AL2.0,LGPL2.1}"
+    }
+  }
+}
+
+dependencies {
+
+  implementation(project(":features:core:calculator"))
+
+  val composeBom = platform(libs.androidx.compose.bom)
+  implementation(composeBom)
+  androidTestImplementation(composeBom)
+
+  implementation(libs.androidx.activity.compose)
+  implementation(libs.androidx.core.ktx)
+  implementation(libs.androidx.core.splashscreen)
+  implementation(libs.androidx.lifecycle.runtime.ktx)
+  implementation(libs.androidx.compose.material.icons.extended)
+  implementation(libs.google.play.services.wearable)
+  implementation(libs.androidx.percentlayout)
+  implementation(libs.androidx.legacy.support.v4)
+  implementation(libs.androidx.recyclerview)
+
+  implementation(libs.androidx.compose.ui)
+  implementation(libs.androidx.compose.ui.tooling.preview)
+  implementation(libs.androidx.wear.compose.material)
+  implementation(libs.androidx.wear.compose.foundation)
+  implementation(libs.androidx.wear.input)
+  implementation(libs.androidx.wear.navigation)
+
+  testImplementation(libs.junit)
+
+  androidTestImplementation(libs.androidx.compose.ui.test.junit4)
+  androidTestImplementation(libs.androidx.test.ext.junit)
+  androidTestImplementation(libs.androidx.test.espresso.core)
+  androidTestImplementation(libs.google.truth)
+
+  debugImplementation(libs.androidx.compose.ui.tooling)
+  debugImplementation(libs.androidx.compose.ui.test.manifest)
+}
