+++ apps/smartwatch-app/proguard-rules.pro
