+++ apps/smartwatch-app/src/main/AndroidManifest.xml
