+++ apps/smartwatch-app/src/main/ic_launcher-playstore.png
