+++ apps/smartwatch-app/src/main/res/drawable/img_splash.webp
