+++ apps/smartwatch-app/src/main/res/drawable/splash_screen.xml
