+++ apps/smartwatch-app/src/main/res/font/comfortaa_bold.ttf
