+++ apps/smartwatch-app/src/main/res/font/comfortaa_medium.ttf
