+++ apps/smartwatch-app/src/main/res/values-es/strings.xml
