+++ apps/smartwatch-app/src/main/res/values/dimens.xml
