+++ apps/smartwatch-app/src/main/res/values/ic_launcher_background.xml
