+++ apps/smartwatch-app/src/main/res/values/strings.xml
