+++ apps/smartwatch-app/src/main/res/values/styles.xml
